requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
outcome==1.2.0
psycopg2-binary==2.9.8
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.5
wsproto==1.2.0
